+++ bindgen/src/bindings/cpp/templates/opt_conv.hpp
@@ -3,5 +3,5 @@ struct {{ typ|ffi_converter_name }} {
     static RustBuffer lower(const {{ type_name }}& val);
     static {{ type_name }} read(RustStream &stream);
     static void write(RustStream &stream, const {{ type_name }}& value);
-    static int32_t allocation_size(const {{ type_name }} &val);
+    static uint64_t allocation_size(const {{ type_name }} &val);
 };
