+++ bindgen/src/bindings/cpp/templates/rec_conv.hpp
@@ -4,5 +4,5 @@ struct {{ typ|ffi_converter_name }} {
     static RustBuffer lower(const {{ class_name }} &);
     static {{ class_name }} read(RustStream &);
     static void write(RustStream &, const {{ class_name }} &);
-    static int32_t allocation_size(const {{ class_name }} &);
+    static uint64_t allocation_size(const {{ class_name }} &);
 };
